DB/미슐랭_크롤링/requirements.txt
playwright==1.49.1
beautifulsoup4==4.12.3
supabase==2.11.0
python-dotenv==1.0.1
tqdm==4.67.1
